+++ src/lib/Components/Consignments/Components/__tests__/FormElements-tests.tsx
@@ -1,0 +1,16 @@
+import * as React from "react"
+import "react-native"
+import * as renderer from "react-test-renderer"
+
+import { Form, Label, Row } from "../FormElements"
+
+describe("Row", () => {
+  it("row passes style props, and other props into the view", () => {
+    const tree = renderer.create(<Row renderToHardwareTextureAndroid={true} style={{ scaleX: 23 }} />).toJSON()
+    expect(tree.props.renderToHardwareTextureAndroid).toBeTruthy()
+
+    const styles = Object.keys(tree.props.style)
+    expect(styles.length).not.toEqual(1)
+    expect(styles).toContain("scaleX")
+  })
+})