+++ src/lib/Components/Consignments/Screens/SelectFromPhotoLibrary.tsx
@@ -23,8 +23,6 @@ import {
   ViewProperties,
 } from "react-native"
 
-import triggerCamera from "../../../NativeModules/triggerCamera"
-
 interface Props extends ViewProperties {
   navigator: NavigatorIOS
   route: Route
@@ -138,14 +136,7 @@ export default class SelectFromPhotoLibrary extends React.Component<Props, State
   }
 
   onPressNewPhoto = () => {
-    triggerCamera(this).then(photo => {
-      if (photo) {
-        console.log("Got photo back")
-        console.log(photo)
-      } else {
-        console.log("Cancelled")
-      }
-    })
+    console.log("OK")
   }
 
   render() {